Migrate calendar from /docs/components/calendar to /components/calendar
Move all examples (Basic, Form, Constraints) from the old documentation
page to the new reference page, remove the old page and its route, and
update E2E tests to target /components/calendar.
